[FEATURE] Filter
Move onchain history to seaprated component
